requirements.txt
openai==1.3.0
SpeechRecognition==3.10.0
PyAudio==0.2.11
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0 